create folder for all routes
