+++ src/Homepage/index.js
@@ -4,7 +4,7 @@ const DEBUG = process.env.NODE_ENV === "development" ? true : false;
 
 const PREFIX = DEBUG ? "http://0.0.0.0:8000" : "";
 
-const api_call = PREFIX + "/hello";
+const api_call = PREFIX + "/api/hello";
 
 const Homepage = () => {
   const [backendData, setBackendData] = useState("");
